requirements.txt
alembic==1.16.4
fastapi==0.116.1
pydantic==2.11.7
python-dotenv==1.1.1
python_jose==3.3.0
SQLAlchemy==2.0.43
uvicorn==0.35.0
